move folders and fix filtering issue
